Refactor parsing of JSON value
